refactor: split git DNS resolution and sink policy into their own modules
